Trait mpi::collective::Root

pub trait Root: AsCommunicator {
    fn root_rank(&self) -> Rank;

    fn broadcast_into<Buf: ?Sized>(&self, buffer: &mut Buf) where Buf: BufferMut { ... }
    fn gather_into<S: ?Sized>(&self, sendbuf: &S) where S: Buffer { ... }
    fn gather_into_root<S: ?Sized, R: ?Sized>(&self,
                                              sendbuf: &S,
                                              recvbuf: &mut R) where S: Buffer, R: BufferMut { ... }
    fn gather_varcount_into<S: ?Sized>(&self, sendbuf: &S) where S: Buffer { ... }
    fn gather_varcount_into_root<S: ?Sized, R: ?Sized>(&self,
                                                       sendbuf: &S,
                                                       recvbuf: &mut R) where S: Buffer, R: PartitionedBufferMut { ... }
    fn scatter_into<R: ?Sized>(&self, recvbuf: &mut R) where R: BufferMut { ... }
    fn scatter_into_root<S: ?Sized, R: ?Sized>(&self,
                                               sendbuf: &S,
                                               recvbuf: &mut R) where S: Buffer, R: BufferMut { ... }
    fn scatter_varcount_into<R: ?Sized>(&self, recvbuf: &mut R) where R: BufferMut { ... }
    fn scatter_varcount_into_root<S: ?Sized, R: ?Sized>(&self,
                                                        sendbuf: &S,
                                                        recvbuf: &mut R) where S: PartitionedBuffer, R: BufferMut { ... }
    fn reduce_into<S: ?Sized, O>(&self, sendbuf: &S, op: &O) where S: Buffer, O: Operation { ... }
    fn reduce_into_root<S: ?Sized, R: ?Sized, O>(&self,
                                                 sendbuf: &S,
                                                 recvbuf: &mut R,
                                                 op: &O) where S: Buffer, R: BufferMut, O: Operation { ... }
    fn immediate_broadcast_into<'b, Buf: ?Sized>(&self,
                                                 buf: &'b mut Buf)
                                                 -> WriteRequest<'b, Buf> where Buf: 'b + BufferMut { ... }
    fn immediate_gather_into<'s, S: ?Sized>(&self,
                                            sendbuf: &'s S)
                                            -> ReadRequest<'s, S> where S: 's + Buffer { ... }
    fn immediate_gather_into_root<'s, 'r, S: ?Sized, R: ?Sized>(&self,
                                                                sendbuf: &'s S,
                                                                recvbuf: &'r mut R)
                                                                -> ReadWriteRequest<'s, 'r, S, R> where S: 's + Buffer, R: 'r + BufferMut { ... }
    fn immediate_gather_varcount_into<'s, S: ?Sized>(&self,
                                                     sendbuf: &'s S)
                                                     -> ReadRequest<'s, S> where S: 's + Buffer { ... }
    fn immediate_gather_varcount_into_root<'s, 'r, S: ?Sized, R: ?Sized>(&self,
                                                                         sendbuf: &'s S,
                                                                         recvbuf: &'r mut R)
                                                                         -> ReadWriteRequest<'s, 'r, S, R> where S: 's + Buffer, R: 'r + PartitionedBufferMut { ... }
    fn immediate_scatter_into<'r, R: ?Sized>(&self,
                                             recvbuf: &'r mut R)
                                             -> WriteRequest<'r, R> where R: 'r + BufferMut { ... }
    fn immediate_scatter_into_root<'s, 'r, S: ?Sized, R: ?Sized>(&self,
                                                                 sendbuf: &'s S,
                                                                 recvbuf: &'r mut R)
                                                                 -> ReadWriteRequest<'s, 'r, S, R> where S: 's + Buffer, R: 'r + BufferMut { ... }
    fn immediate_scatter_varcount_into<'r, R: ?Sized>(&self,
                                                      recvbuf: &'r mut R)
                                                      -> WriteRequest<'r, R> where R: 'r + BufferMut { ... }
    fn immediate_scatter_varcount_into_root<'s, 'r, S: ?Sized, R: ?Sized>(&self,
                                                                          sendbuf: &'s S,
                                                                          recvbuf: &'r mut R)
                                                                          -> ReadWriteRequest<'s, 'r, S, R> where S: 's + PartitionedBuffer, R: 'r + BufferMut { ... }
    fn immediate_reduce_into<'s, S: ?Sized, O>(&self,
                                               sendbuf: &'s S,
                                               op: &O)
                                               -> ReadRequest<'s, S> where S: 's + Buffer, O: Operation { ... }
    fn immediate_reduce_into_root<'s, 'r, S: ?Sized, R: ?Sized, O>(&self,
                                                                   sendbuf: &'s S,
                                                                   recvbuf: &'r mut R,
                                                                   op: &O)
                                                                   -> ReadWriteRequest<'s, 'r, S, R> where S: 's + Buffer, R: 'r + BufferMut, O: Operation { ... }
}

Something that can take the role of 'root' in a collective operation.

Many collective operations define a 'root' process that takes a special role in the communication. These collective operations are implemented as default methods of this trait.

Required Methods

fn root_rank(&self) -> Rank

Rank of the root process

Provided Methods

fn broadcast_into<Buf: ?Sized>(&self, buffer: &mut Buf) where Buf: BufferMut

Broadcast of the contents of a buffer

After the call completes, the Buffer on all processes in the Communicator of the Root &self will contain what it contains on the Root.

Examples

See examples/broadcast.rs

Standard section(s)

5.4

fn gather_into<S: ?Sized>(&self, sendbuf: &S) where S: Buffer

Gather contents of buffers on Root.

After the call completes, the contents of the Buffers on all ranks will be concatenated into the Buffer on Root.

All send Buffers must have the same count of elements.

This function must be called on all non-root processes.

Examples

See examples/gather.rs

Standard section(s)

5.5

fn gather_into_root<S: ?Sized, R: ?Sized>(&self, sendbuf: &S, recvbuf: &mut R) where S: Buffer, R: BufferMut

Gather contents of buffers on Root.

After the call completes, the contents of the Buffers on all ranks will be concatenated into the Buffer on Root.

All send Buffers must have the same count of elements.

This function must be called on the root process.

Examples

See examples/gather.rs

Standard section(s)

5.5

fn gather_varcount_into<S: ?Sized>(&self, sendbuf: &S) where S: Buffer

Gather contents of buffers on Root.

After the call completes, the contents of the Buffers on all ranks will be concatenated into the Buffer on Root.

The send Buffers may contain different counts of elements on different processes. The distribution of elements in the receive Buffer is specified via Partitioned.

This function must be called on all non-root processes.

Examples

See examples/gather_varcount.rs

Standard section(s)

5.5

fn gather_varcount_into_root<S: ?Sized, R: ?Sized>(&self,
sendbuf: &S,
recvbuf: &mut R) where S: Buffer, R: PartitionedBufferMut

Gather contents of buffers on Root.

After the call completes, the contents of the Buffers on all ranks will be concatenated into the Buffer on Root.

The send Buffers may contain different counts of elements on different processes. The distribution of elements in the receive Buffer is specified via Partitioned.

This function must be called on the root process.

Examples

See examples/gather_varcount.rs

Standard section(s)

5.5

fn scatter_into<R: ?Sized>(&self, recvbuf: &mut R) where R: BufferMut

Scatter contents of a buffer on the root process to all processes.

After the call completes each participating process will have received a part of the send Buffer on the root process.

All send Buffers must have the same count of elements.

This function must be called on all non-root processes.

Examples

See examples/scatter.rs

Standard section(s)

5.6

fn scatter_into_root<S: ?Sized, R: ?Sized>(&self, sendbuf: &S, recvbuf: &mut R) where S: Buffer, R: BufferMut

Scatter contents of a buffer on the root process to all processes.

After the call completes each participating process will have received a part of the send Buffer on the root process.

All send Buffers must have the same count of elements.

This function must be called on the root process.

Examples

See examples/scatter.rs

Standard section(s)

5.6

fn scatter_varcount_into<R: ?Sized>(&self, recvbuf: &mut R) where R: BufferMut

Scatter contents of a buffer on the root process to all processes.

After the call completes each participating process will have received a part of the send Buffer on the root process.

The send Buffer may contain different counts of elements for different processes. The distribution of elements in the send Buffer is specified via Partitioned.

This function must be called on all non-root processes.

Examples

See examples/scatter_varcount.rs

Standard section(s)

5.6

fn scatter_varcount_into_root<S: ?Sized, R: ?Sized>(&self,
sendbuf: &S,
recvbuf: &mut R) where S: PartitionedBuffer, R: BufferMut

Scatter contents of a buffer on the root process to all processes.

After the call completes each participating process will have received a part of the send Buffer on the root process.

The send Buffer may contain different counts of elements for different processes. The distribution of elements in the send Buffer is specified via Partitioned.

This function must be called on the root process.

Examples

See examples/scatter_varcount.rs

Standard section(s)

5.6

fn reduce_into<S: ?Sized, O>(&self, sendbuf: &S, op: &O) where S: Buffer, O: Operation

Performs a global reduction under the operation op of the input data in sendbuf and stores the result on the Root process.

This function must be called on all non-root processes.

Examples

See examples/reduce.rs

Standard section(s)

5.9.1

fn reduce_into_root<S: ?Sized, R: ?Sized, O>(&self,
sendbuf: &S,
recvbuf: &mut R,
op: &O) where S: Buffer, R: BufferMut, O: Operation

Performs a global reduction under the operation op of the input data in sendbuf and stores the result on the Root process.

This function must be called on the root process.

Examples

See examples/reduce.rs

Standard section(s)

5.9.1

fn immediate_broadcast_into<'b, Buf: ?Sized>(&self,
buf: &'b mut Buf)
-> WriteRequest<'b, Buf> where Buf: 'b + BufferMut

Initiate broadcast of a value from the Root process to all other processes.

Examples

See examples/immediate_broadcast.rs

Standard section(s)

5.12.2

fn immediate_gather_into<'s, S: ?Sized>(&self,
sendbuf: &'s S)
-> ReadRequest<'s, S> where S: 's + Buffer

Initiate non-blocking gather of the contents of all sendbufs on Root &self.

This function must be called on all non-root processes.

Examples

See examples/immediate_gather.rs

Standard section(s)

5.12.3

fn immediate_gather_into_root<'s, 'r, S: ?Sized, R: ?Sized>(&self,
sendbuf: &'s S,
recvbuf: &'r mut R)
-> ReadWriteRequest<'s, 'r, S, R> where S: 's + Buffer, R: 'r + BufferMut

Initiate non-blocking gather of the contents of all sendbufs on Root &self.

This function must be called on the root processes.

Examples

See examples/immediate_gather.rs

Standard section(s)

5.12.3

fn immediate_gather_varcount_into<'s, S: ?Sized>(&self,
sendbuf: &'s S)
-> ReadRequest<'s, S> where S: 's + Buffer

Initiate non-blocking gather of the contents of all sendbufs on Root &self.

This function must be called on all non-root processes.

Examples

See examples/immediate_gather_varcount.rs

Standard section(s)

5.12.3

fn immediate_gather_varcount_into_root<'s, 'r, S: ?Sized, R: ?Sized>(&self,
sendbuf: &'s S,
recvbuf: &'r mut R)
-> ReadWriteRequest<'s, 'r, S, R> where S: 's + Buffer, R: 'r + PartitionedBufferMut

Initiate non-blocking gather of the contents of all sendbufs on Root &self.

This function must be called on the root processes.

Examples

See examples/immediate_gather_varcount.rs

Standard section(s)

5.12.3

fn immediate_scatter_into<'r, R: ?Sized>(&self,
recvbuf: &'r mut R)
-> WriteRequest<'r, R> where R: 'r + BufferMut

Initiate non-blocking scatter of the contents of sendbuf from Root &self.

This function must be called on all non-root processes.

Examples

See examples/immediate_scatter.rs

Standard section(s)

5.12.4

fn immediate_scatter_into_root<'s, 'r, S: ?Sized, R: ?Sized>(&self,
sendbuf: &'s S,
recvbuf: &'r mut R)
-> ReadWriteRequest<'s, 'r, S, R> where S: 's + Buffer, R: 'r + BufferMut

Initiate non-blocking scatter of the contents of sendbuf from Root &self.

This function must be called on the root processes.

Examples

See examples/immediate_scatter.rs

Standard section(s)

5.12.4

fn immediate_scatter_varcount_into<'r, R: ?Sized>(&self,
recvbuf: &'r mut R)
-> WriteRequest<'r, R> where R: 'r + BufferMut

Initiate non-blocking scatter of the contents of sendbuf from Root &self.

This function must be called on all non-root processes.

Examples

See examples/immediate_scatter_varcount.rs

Standard section(s)

5.12.4

fn immediate_scatter_varcount_into_root<'s, 'r, S: ?Sized, R: ?Sized>(&self,
sendbuf: &'s S,
recvbuf: &'r mut R)
-> ReadWriteRequest<'s, 'r, S, R> where S: 's + PartitionedBuffer, R: 'r + BufferMut

Initiate non-blocking scatter of the contents of sendbuf from Root &self.

This function must be called on the root processes.

Examples

See examples/immediate_scatter_varcount.rs

Standard section(s)

5.12.4

fn immediate_reduce_into<'s, S: ?Sized, O>(&self,
sendbuf: &'s S,
op: &O)
-> ReadRequest<'s, S> where S: 's + Buffer, O: Operation

Initiates a non-blacking global reduction under the operation op of the input data in sendbuf and stores the result on the Root process.

This function must be called on all non-root processes.

Examples

See examples/immediate_reduce.rs

Standard section(s)

5.12.7

fn immediate_reduce_into_root<'s, 'r, S: ?Sized, R: ?Sized, O>(&self,
sendbuf: &'s S,
recvbuf: &'r mut R,
op: &O)
-> ReadWriteRequest<'s, 'r, S, R> where S: 's + Buffer, R: 'r + BufferMut, O: Operation

Initiates a non-blocking global reduction under the operation op of the input data in sendbuf and stores the result on the Root process.

Examples

See examples/immediate_reduce.rs

This function must be called on the root process.

Standard section(s)

5.12.7