Trait mpi::collective::ScatterInto

pub trait ScatterInto {
    fn scatter_into<R: BufferMut + ?Sized>(&self, recvbuf: &mut R);
    fn scatter_into_root<S: Buffer + ?Sized, R: BufferMut + ?Sized>(&self, sendbuf: &S, recvbuf: &mut R);
}

Scatter contents of a buffer on the root process to all processes.

After the call completes each participating process will have received a part of the send Buffer on the root process.

Standard section(s)

5.6

Required Methods

fn scatter_into<R: BufferMut + ?Sized>(&self, recvbuf: &mut R)

Scatter data from the root process to all participating processes.

This function must be called on all non-root processes.

Examples

See examples/scatter.rs

fn scatter_into_root<S: Buffer + ?Sized, R: BufferMut + ?Sized>(&self, sendbuf: &S, recvbuf: &mut R)

Scatter the contents of sendbuf to all participating processes.

This function must be called on the root process.

Examples

See examples/scatter.rs