Struct mpi::datatype::DynBuffer

pub struct DynBuffer<'a> { /* fields omitted */ }

An immutable dynamically-typed buffer.

The buffer has a definite length and MPI datatype, but it is not yet known which Rust type it corresponds to. This is the MPI analogue of &Any. It is semantically equivalent to the trait object reference &Buffer.

Methods

impl<'a> DynBuffer<'a>

pub fn new<T: Equivalence>(buf: &'a [T]) -> Self

Creates a buffer from a slice with whose type has an MPI equivalent.

pub fn is<T: Equivalence>(&self) -> bool

Tests whether the buffer type matches T.

pub fn downcast<T: Equivalence>(self) -> Option<&'a [T]>

Returns some slice if the type matches T, or None if it doesn't.

pub unsafe fn from_raw<T>(
ptr: *const T,
len: Count,
datatype: DatatypeRef<'a>
) -> Self

Creates a buffer from its raw components. The buffer must remain valid for 'a and the pointer must not be null.

pub fn len(&self) -> usize

Returns the number of elements in the buffer.

pub fn is_empty(&self) -> bool

Returns true if the buffer is empty

pub fn as_ptr(&self) -> *const c_void

Returns the underlying raw pointer.