Struct mpi::datatype::DynBufferMut

pub struct DynBufferMut<'a> { /* fields omitted */ }

A mutable dynamically-typed buffer.

The buffer has a definite length and MPI datatype, but it is not yet known which Rust type it corresponds to. This is the MPI analogue of &mut Any. It is semantically equivalent to the mutable trait object reference &mut BufferMut.

Methods

impl<'a> DynBufferMut<'a>

pub fn new<T: Equivalence>(buf: &'a mut [T]) -> Self

Creates a mutable buffer from a mutable slice with whose type has an MPI equivalent.

pub fn is<T: Equivalence>(&self) -> bool

Tests whether the buffer type matches T.

pub fn downcast<T: Equivalence>(self) -> Option<&'a mut [T]>

Returns some mutable slice if the type matches T, or None if it doesn't.

pub unsafe fn from_raw<T>(
ptr: *mut T,
len: Count,
datatype: DatatypeRef<'a>
) -> Self

Creates a buffer from its raw components. The buffer must remain valid for 'a and the pointer must not be null.

pub fn len(&self) -> usize

Returns the number of elements in the buffer.

pub fn is_empty(&self) -> bool

Returns true if the buffer is empty

pub fn as_ptr(&self) -> *const c_void

Returns the underlying raw pointer.

pub fn as_mut_ptr(&mut self) -> *mut c_void

Returns the underlying raw pointer.

pub fn reborrow(&self) -> DynBuffer

Reborrows the buffer with a shorter lifetime.

pub fn reborrow_mut(&mut self) -> DynBufferMut

Reborrows the buffer mutably with a shorter lifetime.

pub fn downgrade(self) -> DynBuffer<'a>

Makes the buffer immutable.