Trait mpi::datatype::Equivalence

pub unsafe trait Equivalence {
    type Out: Datatype;
    fn equivalent_datatype() -> Self::Out;
}

A direct equivalence exists between the implementing type and an MPI datatype

Standard section(s)

3.2.2

Associated Types

type Out: Datatype

The type of the equivalent MPI datatype (e.g. SystemDatatype or UserDatatype)

Required Methods

fn equivalent_datatype() -> Self::Out

The MPI datatype that is equivalent to this Rust type