Trait mpi::datatype::UncommittedDatatype

pub trait UncommittedDatatype: AsRaw<Raw = MPI_Datatype> {
    type DuplicatedDatatype: FromRaw<Raw = MPI_Datatype>;
    fn dup(&self) -> Self::DuplicatedDatatype { ... }
}

An UncommittedDatatype is a partial description of the layout of messages in memory which may not yet have been committed to an implementation-defined message format.

Committed datatypes can be treated as-if they are uncommitted.

Associated Types

type DuplicatedDatatype: FromRaw<Raw = MPI_Datatype>

The type returned when the datatype is duplicated.

Loading content...
Provided methods

fn dup(&self) -> Self::DuplicatedDatatype

Creates a new datatype with the same key-values as this datatype.

Standard section(s)

4.1.10

Loading content...