Function mpi::initialize

pub fn initialize() -> Option<Universe>

Initialize MPI.

If the MPI library has not been initialized so far, initializes and returns a representation of the MPI communication Universe which provides access to additional functions. Otherwise returns None.

Equivalent to: initialize_with_threading(Threading::Single)

Examples

See examples/simple.rs

Standard section(s)

8.7