Function mpi::initialize_with_threading

pub fn initialize_with_threading(
    threading: Threading
) -> Option<(Universe, Threading)>

Initialize MPI with desired level of multithreading support.

If the MPI library has not been initialized so far, tries to initialize with the desired level of multithreading support and returns the MPI communication Universe with access to additional functions as well as the level of multithreading actually supported by the implementation. Otherwise returns None.

Examples

See examples/init_with_threading.rs

Standard section(s)

12.4.3