Function mpi::point_to_point::send_receive_replace_into

pub fn send_receive_replace_into<B: ?Sized, D, S>(
    buf: &mut B, 
    destination: &D, 
    source: &S
) -> Status where
    B: BufferMut,
    D: Destination,
    S: Source, 

Sends the contents of buf to destination and simultaneously receives a message from source and replaces the contents of buf with it.

Standard section(s)

3.10