Trait mpi::point_to_point::MatchedReceiveVec

pub trait MatchedReceiveVec {
    fn matched_receive_vec<Msg>(self) -> (Vec<Msg>, Status)
    where
        Msg: Equivalence;
}

Receive a previously probed message containing multiple instances of type Msg into a Vec.

Standard section(s)

3.8.3

Required Methods

fn matched_receive_vec<Msg>(self) -> (Vec<Msg>, Status) where
Msg: Equivalence,

Receives the message &self which contains multiple instances of type Msg into a Vec.