Trait mpi::point_to_point::ReadySend

pub trait ReadySend {
    fn ready_send_with_tag<Buf: Buffer + ?Sized>(&self, buf: &Buf, tag: Tag);

    fn ready_send<Buf: Buffer + ?Sized>(&self, buf: &Buf) { ... }
}

Blocking ready mode send operation

Standard section(s)

3.4

Required Methods

fn ready_send_with_tag<Buf: Buffer + ?Sized>(&self, buf: &Buf, tag: Tag)

Send the contents of a Buffer to the Destination &self and tag it.

Fails if the matching receive operation has not been posted.

Provided Methods

fn ready_send<Buf: Buffer + ?Sized>(&self, buf: &Buf)

Send the contents of a Buffer to the Destination &self.

Fails if the matching receive operation has not been posted.