Trait mpi::point_to_point::Receive

pub trait Receive {
    fn receive_with_tag<Msg: EquivalentDatatype>(&self, tag: Tag) -> (Option<Msg>, Status);

    fn receive<Msg: EquivalentDatatype>(&self) -> (Option<Msg>, Status) { ... }
}

Receive a message containing a single instance of type Msg.

Examples

use mpi::traits::*;

let universe = mpi::initialize().unwrap();
let world = universe.world();

let x = world.receive::<f64>();

Standard section(s)

3.2.4

Required Methods

fn receive_with_tag<Msg: EquivalentDatatype>(&self, tag: Tag) -> (Option<Msg>, Status)

Receive a message from Source &self tagged tag containing a single instance of type Msg or None if receiving from the null process.

Provided Methods

fn receive<Msg: EquivalentDatatype>(&self) -> (Option<Msg>, Status)

Receive a message from Source &self containing a single instance of type Msg or None if receiving from the null process.