Trait mpi::point_to_point::Send

pub trait Send {
    fn send_with_tag<Buf: Buffer + ?Sized>(&self, buf: &Buf, tag: Tag);

    fn send<Buf: Buffer + ?Sized>(&self, buf: &Buf) { ... }
}

Blocking standard mode send operation

Examples

use mpi::traits::*;

let universe = mpi::initialize().unwrap();
let world = universe.world();

let v = vec![ 1.0f64, 2.0, 3.0 ];
world.process_at_rank(1).send(&v[..]);

Standard section(s)

3.2.1

Required Methods

fn send_with_tag<Buf: Buffer + ?Sized>(&self, buf: &Buf, tag: Tag)

Send the contents of a Buffer to the Destination &self and tag it.

Provided Methods

fn send<Buf: Buffer + ?Sized>(&self, buf: &Buf)

Send the contents of a Buffer to the Destination &self.

Examples

See examples/send_receive.rs