Trait mpi::point_to_point::SendReceive

pub trait SendReceive {
    fn send_receive_with_tags<S, R>(&self, msg: &S, destination: Rank, sendtag: Tag, source: Rank, receivetag: Tag) -> (Option<R>, Status) where S: EquivalentDatatype, R: EquivalentDatatype;

    fn send_receive<S, R>(&self, msg: &S, destination: Rank, source: Rank) -> (Option<R>, Status) where S: EquivalentDatatype, R: EquivalentDatatype { ... }
}

Simultaneously send and receive single instances of types S and R.

Standard section(s)

3.10

Required Methods

fn send_receive_with_tags<S, R>(&self, msg: &S, destination: Rank, sendtag: Tag, source: Rank, receivetag: Tag) -> (Option<R>, Status) where S: EquivalentDatatype, R: EquivalentDatatype

Sends msg to Rank destination tagging it sendtag and simultaneously receives an instance of R tagged receivetag from Rank source or receives None if receiving from the null process.

Provided Methods

fn send_receive<S, R>(&self, msg: &S, destination: Rank, source: Rank) -> (Option<R>, Status) where S: EquivalentDatatype, R: EquivalentDatatype

Sends msg to Rank destination and simultaneously receives an instance of R from Rank source or receives None if receiving from the null process.

Examples

See examples/send_receive.rs