Trait mpi::point_to_point::SendReceiveReplaceInto

pub trait SendReceiveReplaceInto {
    fn send_receive_replace_into_with_tags<B: ?Sized>(&self, buf: &mut B, destination: Rank, sendtag: Tag, source: Rank, receivetag: Tag) -> Status where B: BufferMut;

    fn send_receive_replace_into<B: ?Sized>(&self, buf: &mut B, destination: Rank, source: Rank) -> Status where B: BufferMut { ... }
}

Simultaneously send and receive the contents of the same buffer.

Standard section(s)

3.10

Required Methods

fn send_receive_replace_into_with_tags<B: ?Sized>(&self, buf: &mut B, destination: Rank, sendtag: Tag, source: Rank, receivetag: Tag) -> Status where B: BufferMut

Sends the contents of buf to Rank destination tagging it sendtag and simultaneously receives a message tagged receivetag from Rank source and replaces the contents of buf with it.

Receiving from the null process leaves buf untouched.

Provided Methods

fn send_receive_replace_into<B: ?Sized>(&self, buf: &mut B, destination: Rank, source: Rank) -> Status where B: BufferMut

Sends the contents of buf to Rank destination and simultaneously receives a message from Rank source into and replaces the contents of buf with it.

Receiving from the null process leaves buf untouched.