Trait mpi::raw::FromRaw

pub trait FromRaw: AsRaw {
    unsafe fn from_raw(handle: Self::Raw) -> Self;
}

Conversion for the Rust type from the raw MPI handle type.

Required methods

unsafe fn from_raw(handle: Self::Raw) -> Self

Constructs the Rust type, with all its semantics, from the raw MPI type.

Loading content...