Struct mpi::request::CancelGuard

pub struct CancelGuard<'a, S: Scope<'a>>(_);

Guard object that tries to cancel and waits for the completion of an operation when it is dropped

The guard can be constructed or deconstructed using the From and Into traits.

Examples

See examples/immediate.rs