Struct mpi::request::StaticScope

pub struct StaticScope;

The scope that lasts as long as the entire execution of the program

Unlike LocalScope<'a>, StaticScope does not require any bookkeeping on the requests as every request associated with a StaticScope can live as long as they please.

A StaticScope can be created simply by calling the StaticScope constructor.

Invariant

For any Request registered with a StaticScope, its associated buffers must be 'static.