Struct mpi::request::WaitGuard

pub struct WaitGuard<'a, S: Scope<'a> = StaticScope>(_);

Guard object that waits for the completion of an operation when it is dropped

The guard can be constructed or deconstructed using the From and Into traits.

Examples

See examples/immediate.rs