Trait mpi::request::Request

pub trait Request: AsRaw<Raw = MPI_Request> + AsRawMut {
    fn is_null(&self) -> bool { ... }
    fn wait(self) -> Status where Self: Sized { ... }
    fn test(self) -> Result<Status, Self> where Self: Sized { ... }
    fn cancel(self) where Self: Sized { ... }
}

A request for a non-blocking operation

Provided Methods

fn is_null(&self) -> bool

Returns true for a null request handle.

fn wait(self) -> Status where Self: Sized

Wait for an operation to finish.

Will block execution of the calling thread until the associated operation has finished.

Examples

See examples/immediate.rs

Standard section(s)

3.7.3

fn test(self) -> Result<Status, Self> where Self: Sized

Test whether an operation has finished.

If the operation has finished returns the Status otherwise returns the unfinished Request.

Examples

See examples/immediate.rs

Standard section(s)

3.7.3

fn cancel(self) where Self: Sized

Cancel an operation.

Examples

See examples/immediate.rs

Standard section(s)

3.8.4