Struct mpi::topology::CartesianCommunicator

pub struct CartesianCommunicator(_);

A CartesianCommunicator is an MPI communicator object where ranks are laid out in an n-dimensional cartesian space. This gives ranks neighbors in each of those dimensions, and MPI is able to optimize the layout of these ranks to improve physical locality.

Standard Section(s)

7

Implementations

impl CartesianCommunicator

pub unsafe fn from_raw(raw: MPI_Comm) -> Option<CartesianCommunicator>

Given a valid MPI_Comm handle in raw, returns a CartesianCommunicator value if, and only if:

The handle is not MPI_COMM_NULL

The topology of the communicator is MPI_CART

Otherwise returns None.

Parameters

raw - Handle to a valid MPI_Comm object

Safety

raw must be a live MPI_Comm object.

raw must not be used after calling from_raw.

pub unsafe fn from_raw_unchecked(raw: MPI_Comm) -> CartesianCommunicator

Creates a CartesianCommunicator from raw.

Parameters

raw - Handle to a valid MPI_CART MPI_Comm object

Safety

raw must be a live MPI_Comm object.

raw must not be used after calling from_raw_unchecked.

raw must not be MPI_COMM_NULL.

pub fn num_dimensions(&self) -> Count

Returns the number of dimensions that the Cartesian communicator was established over.

Standard section(s)

7.5.5 (MPI_Cartdim_get)

pub unsafe fn get_layout_into_unchecked( &self, dims: &mut [Count], periods: &mut [bool], coords: &mut [Count] )

Returns the topological structure of the Cartesian communicator

Prefer get_layout_into

Parameters

dims - array of spatial extents for the cartesian space

periods - Must match length of dims. periods[i] indicates if axis i is periodic. i.e. if true, the element at dims[i] - 1 in axis i is a neighbor of element 0 in axis i

coords - coords[i] is the location offset in axis i of this rank

Standard section(s)

7.5.5 (MPI_Cart_get)

Safety

Behavior is undefined if dims, periods, and coords are not of length num_dimensions.

pub fn get_layout_into( &self, dims: &mut [Count], periods: &mut [bool], coords: &mut [Count] )

Returns the topological structure of the Cartesian communicator

Panics if dims, periods, and coords are not of length num_dimensions.

Parameters

dims - array of spatial extents for the cartesian space

periods - Must match length of dims. periods[i] indicates if axis i is periodic. i.e. if true, the element at dims[i] - 1 in axis i is a neighbor of element 0 in axis i

coords - coords[i] is the location offset in axis i of this rank

Standard section(s)

7.5.5 (MPI_Cart_get)

pub fn get_layout(&self) -> CartesianLayout

Returns the topological structure of the Cartesian communicator

Standard section(s)

7.5.5 (MPI_Cart_get)

pub unsafe fn coordinates_to_rank_unchecked(&self, coords: &[Count]) -> Rank

Converts a set of cartesian coordinates to its rank in the CartesianCommunicator.

Coordinates in periodic axes that are out of range are shifted back into the dimensions of the communiactor.

Prefer coordinates_to_rank

Parameters

coords - coords[i] is a location offset in axis i

Standard section(s)

7.5.5 (MPI_Cart_rank)

Safety

Behavior is undefined if coords is not of length num_dimensions.

Behavior is undefined if any coordinates in non-periodic axes are outside the dimensions

pub fn coordinates_to_rank(&self, coords: &[Count]) -> Rank

Converts a set of cartesian coordinates to its rank in the CartesianCommunicator.

Panics if coords is not of length num_dimensions.

Coordinates in periodic axes that are out of range are shifted back into the dimensions of the communiactor.

Panics if any coordinates in non-periodic axes are outside the dimensions of the communicator.

Parameters

coords - coords[i] is a location offset in axis i

Standard section(s)

7.5.5 (MPI_Cart_rank)

pub unsafe fn rank_to_coordinates_into_unchecked( &self, rank: Rank, coords: &mut [Count] )

Receives into coords the cartesian coordinates of rank.

Prefer rank_to_coordinates_into

Parameters

rank - A rank in the communicator

coords - coords[i] is the cartesian coordinate of rank in axis i

Standard section(s)

7.5.5 (MPI_Cart_coords)

Safety

Behavior is undefined if rank is not a non-negative value less than size.

pub fn rank_to_coordinates_into(&self, rank: Rank, coords: &mut [Count])

Receives into coords the cartesian coordinates of rank.

Panics if rank is not a non-negative value less than size.

Parameters

rank - A rank in the communicator

coords - coords[i] is the cartesian coordinate of rank in axis i

Standard section(s)

7.5.5 (MPI_Cart_coords)

pub fn rank_to_coordinates(&self, rank: Rank) -> Vec<Count>

Returns an array of coords with the cartesian coordinates of rank, where coords[i] is the cartesian coordinate of rank in axis i.

Panics if rank is not a non-negative value less than size.

Parameters

rank - A rank in the communicator

Standard section(s)

7.5.5 (MPI_Cart_coords)

pub unsafe fn shift_unchecked( &self, dimension: Count, displacement: Count ) -> (Option<Rank>, Option<Rank>)

Retrieves targets in dimension shifted from the current rank by displacing in the negative direction by displacement units for the first returned rank and in the positive direction for the second returned rank.

Prefer shift

Parameters

dimension - which axis to shift in

displacement - what offset to shift by in each direction

Standard section(s)

7.5.6 (MPI_Cart_shift)

Safety

Behavior is undefined if dimension is not of length num_dimensions.

pub fn shift( &self, dimension: Count, displacement: Count ) -> (Option<Rank>, Option<Rank>)

Retrieves targets in dimension shifted from the current rank by displacing in the negative direction by displacement units for the first returned rank and in the positive direction for the second returned rank.

Panics if dimension is not of length num_dimensions.

Parameters

dimension - which axis to shift in

displacement - what offset to shift by in each direction

Standard section(s)

7.5.6 (MPI_Cart_shift)

pub unsafe fn subgroup_unchecked( &self, retain: &[bool] ) -> CartesianCommunicator

Partitions an existing Cartesian communicator into a new Cartesian communicator in a lower dimension.

Prefer subgroup

Parameters

retain - if retain[i] is true, then axis i is retained in the new communicator

Standard section(s)

7.5.7 (MPI_Cart_sub)

Safety

Behavior is undefined if retain is not of length num_dimensions.

pub fn subgroup(&self, retain: &[bool]) -> CartesianCommunicator

Partitions an existing Cartesian communicator into a new Cartesian communicator in a lower dimension.

Panics if retain is not of length num_dimensions.

Parameters

retain - if retain[i] is true, then axis i is retained in the new communicator

Standard section(s)

7.5.7 (MPI_Cart_sub)