Struct mpi::topology::Universe

pub struct Universe(_);

Global context

Methods

impl Universe

fn world(&self) -> SystemCommunicator

The 'world communicator'

Contains all processes initially partaking in the computation.

Examples

See examples/simple.rs

fn threading_support(&self) -> Threading

Level of multithreading supported by this MPI universe

See the Threading enum.

Examples

See examples/init_with_threading.rs

fn get_processor_name(&self) -> Result<String, FromUtf8Error>

Names the processor that the calling process is running on.

Can return an Err if the processor name is not a UTF-8 string.

fn get_time(&self) -> c_double

Time in seconds since an arbitrary time in the past.

The cheapest high-resolution timer available will be used.

fn get_time_res(&self) -> c_double

Resolution of timer used in MPI_Wtime in seconds