Crate mpi

Modules

bindings
The types and functions defined here are not meant to be used directly. Instead, wrappers are provided in other mods. Ex. prefer to use mpi::comm::rank() instead of unsafe { MPI_Comm_Rank() }.

comm

error
Encapsulates an MPI erorr codes such as MPI_SUCCESS. An Error should only be constructed when Err is returned.
When MPI_SUCCESS is recieved from an MPI_* fn, Ok should be returned.
If an MPI_* fn returns only MPI_SUCCESS then, Result<T, Error> is not needed.

request

status

Functions

finalize
Cleans up mpi stuff. Should always return success, no need for Error.

Panics

When called multiple times. When called without init.

Examples

// Valid
init_null();
finalize();

init_null
Calls MPI_Init with null parameters. As of MPI-2, MPI_Init will accept NULL as input parameters. Must be called before any other mpi function is called.

Panics

When called multiple times. When called without finalize at end of program.

Examples

// Valid
init_null();
finalize();