Module mpi::collective

Collective communication

Developing...

Unfinished features

5.8: All-to-all, MPI_Alltoallw()

5.10: Reduce-scatter, MPI_Reduce_scatter()

5.12: Nonblocking collective operations, MPI_Ialltoallw(), MPI_Ireduce_scatter()

Modules

traits
Collective communication traits

Structs

SystemOperation
A built-in operation like MPI_SUM

UnsafeUserOperation
An unsafe user-defined operation.

UserOperation
A user-defined operation.

Traits

CommunicatorCollectives
Collective communication patterns defined on Communicators

Operation
An operation to be used in a reduction or scan type operation, e.g. MPI_SUM

Root
Something that can take the role of 'root' in a collective operation.

Functions

reduce_local_into
Perform a local reduction.

Type Definitions

UnsafeUserFunction
A raw pointer to a function that can be used to define an UnsafeUserOperation.