Trait mpi::collective::Root

pub trait Root: Communicator {
    fn root_rank(&self) -> Rank;
}

Something that can take the role of 'root' in a collective operation.

Many collective operations define a 'root' process that takes a special role in the communication. These collective operations are implemented as traits that have blanket implementations for every type that implements the Root trait.

Required Methods

fn root_rank(&self) -> Rank

Rank of the root process