Struct mpi::datatype::MutView

pub struct MutView<'d, 'b, D, B: ?Sized> where D: 'd + Datatype,
        B: 'b + PointerMut { /* fields omitted */ }

A buffer with a user specified count and datatype

Safety

Views can be used to instruct the underlying MPI library to rummage around at arbitrary locations in memory. This might be controlled later on using datatype bounds an slice lengths but for now, all View constructors are marked unsafe.

Methods

impl<'d, 'b, D, B: ?Sized> MutView<'d, 'b, D, B> where D: 'd + Datatype,
B: 'b + PointerMut

unsafe fn with_count_and_datatype(buffer: &'b mut B,
count: Count,
datatype: &'d D)
-> MutView<'d, 'b, D, B>

Return a view of buffer containing count instances of MPI datatype datatype.

Examples

See examples/contiguous.rs, examples/vector.rs