Trait mpi::datatype::Partitioned

pub trait Partitioned {
    fn counts(&self) -> &[Count];

    fn displs(&self) -> &[Count];

    unsafe fn counts_ptr(&self) -> *const Count { ... }

    unsafe fn displs_ptr(&self) -> *const Count { ... }
}

Describes how a Buffer is partitioned by specifying the count of elements and displacement from the start of the buffer for each partition.

Required methods

fn counts(&self) -> &[Count]

The count of elements in each partition.

fn displs(&self) -> &[Count]

The displacement from the start of the buffer for each partition.

Loading content...
Provided methods

unsafe fn counts_ptr(&self) -> *const Count

A pointer to counts()

unsafe fn displs_ptr(&self) -> *const Count

A pointer to displs()

Loading content...