Trait mpi::datatype::Partitioned

pub trait Partitioned {
    // Required methods
    fn counts(&self) -> &[Count];
    fn displs(&self) -> &[Count];
}

Describes how a Buffer is partitioned by specifying the count of elements and displacement from the start of the buffer for each partition.

Required Methods

fn counts(&self) -> &[Count]

The count of elements in each partition.

fn displs(&self) -> &[Count]

The displacement from the start of the buffer for each partition.