Module mpi::environment

Environmental management

This module provides ways for an MPI program to interact with its environment.

Unfinished features

8.1.2: MPI_TAG_UB, ...

8.2: Memory allocation

8.3, 8.4, and 8.5: Error handling

Structs

Universe
Global context

Enums

Threading
Describes the various levels of multithreading that can be supported by an MPI library.

Functions

initialize
Initialize MPI.

initialize_with_threading
Initialize MPI with desired level of multithreading support.

library_version
Describes the version of the MPI library itself.

processor_name
Names the processor that the calling process is running on.

threading_support
Level of multithreading supported by this MPI universe

time
Time in seconds since an arbitrary time in the past.

time_resolution
Resolution of timer used in time() in seconds

version
Identifies the version of the MPI standard implemented by the library.