Enum mpi::environment::Threading

pub enum Threading {
    Single,
    Funneled,
    Serialized,
    Multiple,
}

Describes the various levels of multithreading that can be supported by an MPI library.

Examples

See examples/init_with_threading.rs

Standard section(s)

12.4.3

Variants
Single

All processes partaking in the computation are single-threaded.

Funneled

Processes may be multi-threaded, but MPI functions will only ever be called from the main thread.

Serialized

Processes may be multi-threaded, but calls to MPI functions will not be made concurrently. The user is responsible for serializing the calls.

Multiple

Processes may be multi-threaded with no restrictions on the use of MPI functions from the threads.