Module mpi::point_to_point

Point to point communication

Endpoints of communication are mostly described by types that implement the Source and Destination trait. Communication operations are implemented as traits that have blanket implementations for these two traits.

Unfinished features

3.2.6: MPI_STATUS_IGNORE

3.6: Buffer usage, MPI_Buffer_attach(), MPI_Buffer_detach()

3.7: Nonblocking mode:

Completion, MPI_Waitany(), MPI_Waitall(), MPI_Waitsome(), MPI_Testany(), MPI_Testall(), MPI_Testsome(), MPI_Request_get_status()

3.8:

Cancellation, MPI_Cancel(), MPI_Test_cancelled()

3.9: Persistent requests, MPI_Send_init(), MPI_Bsend_init(), MPI_Ssend_init(), MPI_Rsend_init(), MPI_Recv_init(), MPI_Start(), MPI_Startall()

Modules

traits
Point to point communication traits

Structs

Message
Describes a pending incoming message, probed by a matched_probe().

ReceiveFuture
Will contain a value of type T received via a non-blocking receive operation.

ReceiveRequest
A request object for an immediate (non-blocking) receive operation

SendRequest
A request object for an immediate (non-blocking) send operation

Status
Describes the result of a point to point receive operation.

Traits

BufferedSend
Blocking buffered mode send operation

Destination
Something that can be used as the destination in a point to point send operation

ImmediateBufferedSend
Initiate an immediate (non-blocking) buffered mode send operation.

ImmediateMatchedProbe
Asynchronously probe a source for incoming messages with guaranteed reception.

ImmediateMatchedReceiveInto
Asynchronously receive a previously probed message into a Buffer.

ImmediateProbe
Asynchronously probe a source for incoming messages.

ImmediateReadySend
Initiate an immediate (non-blocking) ready mode send operation.

ImmediateReceive
Initiate a non-blocking receive operation.

ImmediateReceiveInto
Initiate an immediate (non-blocking) receive operation.

ImmediateSend
Initiate an immediate (non-blocking) standard mode send operation.

ImmediateSynchronousSend
Initiate an immediate (non-blocking) synchronous mode send operation.

MatchedProbe
Probe a source for incoming messages with guaranteed reception.

MatchedReceive
Receive a previously probed message containing a single instance of type Msg.

MatchedReceiveInto
Receive a previously probed message into a Buffer.

MatchedReceiveVec
Receive a previously probed message containing multiple instances of type Msg into a Vec.

Probe
Probe a source for incoming messages.

ReadySend
Blocking ready mode send operation

Receive
Receive a message containing a single instance of type Msg.

ReceiveInto
Receive a message into a Buffer.

ReceiveVec
Receive a message containing multiple instances of type Msg into a Vec.

Send
Blocking standard mode send operation

SendReceive
Simultaneously send and receive single instances of types S and R.

SendReceiveInto
Simultaneously send and receive the contents of buffers.

SendReceiveReplaceInto
Simultaneously send and receive the contents of the same buffer.

Source
Something that can be used as the source in a point to point receive operation

SynchronousSend
Blocking synchronous mode send operation

Test
Test whether an operation has finished.

Wait
Wait for an operation to finish.