Function mpi::point_to_point::send_receive_with_tags

pub fn send_receive_with_tags<M, D, R, S>(
    msg: &M, 
    destination: &D, 
    sendtag: Tag, 
    source: &S, 
    receivetag: Tag
) -> (R, Status) where
    M: Equivalence,
    D: Destination,
    R: Equivalence,
    S: Source, 

Sends msg to destination tagging it sendtag and simultaneously receives an instance of R tagged receivetag from source.

Standard section(s)

3.10