Struct mpi::point_to_point::Message

#[must_use]
pub struct Message(_);

Describes an pending incoming message, probed by a matched_probe().

Standard section(s)

3.8.2

Methods

impl Message

fn is_no_proc(&self) -> bool

True if the Source for the probe was the null process.