Struct mpi::point_to_point::Message

#[must_use]

pub struct Message(_);

Describes a pending incoming message, probed by a matched_probe().

Standard section(s)

3.8.2

Methods

impl Message

pub fn is_no_proc(&self) -> bool

True if the Source for the probe was the null process.

pub fn matched_receive<Msg>(self) -> (Msg, Status) where
Msg: Equivalence,

Receive a previously probed message containing a single instance of type Msg.

Receives the message &self which contains a single instance of type Msg.

Standard section(s)

3.8.3

pub fn matched_receive_into<Buf: ?Sized>(self, buf: &mut Buf) -> Status where
Buf: BufferMut,

Receive a previously probed message into a Buffer.

Receive the message &self with contents matching buf.

Standard section(s)

3.8.3

pub fn immediate_matched_receive_into<'a, Sc, Buf: ?Sized + 'a>(
self,
scope: Sc,
buf: &'a mut Buf
) -> Request<'a, Sc> where
Buf: BufferMut,
Sc: Scope<'a>,

Asynchronously receive a previously probed message into a Buffer.

Asynchronously receive the message &self with contents matching buf.

Standard section(s)

3.8.3