Trait mpi::raw::AsRawMut

pub unsafe trait AsRawMut: AsRaw {
    fn as_raw_mut(&mut self) -> *mut Self::Raw;
}

A rust type than can provide a mutable pointer to a raw value understood by the MPI C API.

Required Methods

fn as_raw_mut(&mut self) -> *mut Self::Raw

A mutable pointer to the raw value