Trait mpi::raw::FromRaw

pub trait FromRaw: AsRaw {
    // Required method
    unsafe fn from_raw(handle: <Self as AsRaw>::Raw) -> Self;
}

Conversion for the Rust type from the raw MPI handle type.

Required Methods

unsafe fn from_raw(handle: <Self as AsRaw>::Raw) -> Self

Constructs the Rust type, with all its semantics, from the raw MPI type.

Safety

handle may be assumed to be a live MPI object.