Struct mpi::topology::AnyProcess

pub struct AnyProcess<'a, C> where C: 'a + Communicator(_);

Identifies an arbitrary process that is a member of a certain communicator, e.g. for use as a Source in point to point communication.