Struct mpi::topology::AnyProcess

pub struct AnyProcess<'a, C>(_)
 where
    C: 'a + Communicator;

Identifies an arbitrary process that is a member of a certain communicator, e.g. for use as a Source in point to point communication.