Trait mpi::topology::Communicator

pub trait Communicator: AsRaw<Raw = MPI_Comm> {

    // Required method
    fn target_size(&self) -> Rank;

    // Provided methods
    fn size(&self) -> Rank { ... }
    fn rank(&self) -> Rank { ... }
    fn get_attr<A: Attribute>(
        &self,
        key: A
    ) -> Option<&<A as Attribute>::Target>
       where Self: Sized { ... }
    fn process_at_rank(&self, r: Rank) -> Process<'_, Self>
       where Self: Sized { ... }
    fn any_process(&self) -> AnyProcess<'_, Self>
       where Self: Sized { ... }
    fn this_process(&self) -> Process<'_, Self>
       where Self: Sized { ... }
    fn compare<C>(&self, other: &C) -> CommunicatorRelation
       where C: Communicator + ?Sized,
             Self: Sized { ... }
    fn duplicate(&self) -> UserCommunicator { ... }
    fn split_by_color(&self, color: Color) -> Option<UserCommunicator> { ... }
    fn split_by_color_with_key(
        &self,
        color: Color,
        key: Key
    ) -> Option<UserCommunicator> { ... }
    fn split_shared(&self, key: c_int) -> UserCommunicator { ... }
    fn split_by_subgroup_collective<G>(
        &self,
        group: &G
    ) -> Option<UserCommunicator>
       where G: Group + ?Sized,
             Self: Sized { ... }
    fn split_by_subgroup<G>(&self, group: &G) -> Option<UserCommunicator>
       where G: Group + ?Sized,
             Self: Sized { ... }
    fn split_by_subgroup_with_tag<G>(
        &self,
        group: &G,
        tag: Tag
    ) -> Option<UserCommunicator>
       where G: Group + ?Sized,
             Self: Sized { ... }
    fn group(&self) -> UserGroup { ... }
    fn abort(&self, errorcode: c_int) -> ! { ... }
    fn test_inter(&self) -> bool { ... }
    fn set_name(&self, name: &str) { ... }
    fn get_name(&self) -> String { ... }
    fn create_cartesian_communicator(
        &self,
        dims: &[Count],
        periods: &[bool],
        reorder: bool
    ) -> Option<CartesianCommunicator> { ... }
    fn cartesian_map(&self, dims: &[Count], periods: &[bool]) -> Option<Rank> { ... }
    fn pack_size<Dt>(&self, incount: Count, datatype: &Dt) -> Count
       where Dt: Datatype,
             Self: Sized { ... }
    fn pack<Buf>(&self, inbuf: &Buf) -> Vec<u8>
       where Buf: ?Sized + Buffer,
             Self: Sized { ... }
    fn pack_into<Buf>(
        &self,
        inbuf: &Buf,
        outbuf: &mut [u8],
        position: Count
    ) -> Count
       where Buf: ?Sized + Buffer,
             Self: Sized { ... }
    unsafe fn unpack_into<Buf>(
        &self,
        inbuf: &[u8],
        outbuf: &mut Buf,
        position: Count
    ) -> Count
       where Buf: ?Sized + BufferMut,
             Self: Sized { ... }
    fn parent(&self) -> Option<SystemInterCommunicator> { ... }

}

Communicators are contexts for communication

Required Methods

fn target_size(&self) -> Rank

Returns the number of processes available to communicate with in this Communicator. For intra-communicators, this is equivalent to size. For inter-communicators, this is equivalent to remote_size.

Provided Methods

fn size(&self) -> Rank

Number of processes in this communicator

Examples

See examples/simple.rs

Standard section(s)

6.4.1

fn rank(&self) -> Rank

The Rank that identifies the calling process within this communicator

Examples

See examples/simple.rs

Standard section(s)

6.4.1

fn get_attr<A: Attribute>(&self, key: A) -> Option<&<A as Attribute>::Target>where Self: Sized,

Get Attribute

fn process_at_rank(&self, r: Rank) -> Process<'_, Self>where Self: Sized,

Bundles a reference to this communicator with a specific Rank into a Process.

Examples

See examples/broadcast.rs examples/gather.rs examples/send_receive.rs

fn any_process(&self) -> AnyProcess<'_, Self>where Self: Sized,

Returns an AnyProcess identifier that can be used, e.g. as a Source in point to point communication.

fn this_process(&self) -> Process<'_, Self>where Self: Sized,

A Process for the calling process

fn compare<C>(&self, other: &C) -> CommunicatorRelationwhere C: Communicator + ?Sized, Self: Sized,

Compare two communicators.

See enum CommunicatorRelation.

Standard section(s)

6.4.1

fn duplicate(&self) -> UserCommunicator

Duplicate a communicator.

Examples

See examples/duplicate.rs

Standard section(s)

6.4.2

fn split_by_color(&self, color: Color) -> Option<UserCommunicator>

Split a communicator by color.

Creates as many new communicators as distinct values of color are given. All processes with the same value of color join the same communicator. A process that passes the special undefined color will not join a new communicator and None is returned.

Examples

See examples/split.rs

Standard section(s)

6.4.2

fn split_by_color_with_key( &self, color: Color, key: Key ) -> Option<UserCommunicator>

Split a communicator by color.

Like split() but orders processes according to the value of key in the new communicators.

Standard section(s)

6.4.2

fn split_shared(&self, key: c_int) -> UserCommunicator

Split the communicator into subcommunicators, each of which can create a shared memory region.

Within each subgroup, the processes are ranked in the order defined by the value of the argument key, with ties broken according to their rank in the old group.

Standard section(s)

6.4.2 (See: MPI_Comm_split_type)

fn split_by_subgroup_collective<G>(&self, group: &G) -> Option<UserCommunicator>where G: Group + ?Sized, Self: Sized,

Split a communicator collectively by subgroup.

Proceses pass in a group that is a subgroup of the group associated with the old communicator. Different processes may pass in different groups, but if two groups are different, they have to be disjunct. One new communicator is created for each distinct group. The new communicator is returned if a process is a member of the group he passed in, otherwise None.

This call is a collective operation on the old communicator so all processes have to partake.

Examples

See examples/split.rs

Standard section(s)

6.4.2

fn split_by_subgroup<G>(&self, group: &G) -> Option<UserCommunicator>where G: Group + ?Sized, Self: Sized,

Split a communicator by subgroup.

Like split_by_subgroup_collective() but not a collective operation.

Examples

See examples/split.rs

Standard section(s)

6.4.2

fn split_by_subgroup_with_tag<G>( &self, group: &G, tag: Tag ) -> Option<UserCommunicator>where G: Group + ?Sized, Self: Sized,

Split a communicator by subgroup

Like split_by_subgroup() but can avoid collision of concurrent calls (i.e. multithreaded) by passing in distinct tags.

Standard section(s)

6.4.2

fn group(&self) -> UserGroup

The group associated with this communicator

Standard section(s)

6.3.2

fn abort(&self, errorcode: c_int) -> !

Abort program execution

Standard section(s)

8.7

fn test_inter(&self) -> bool

Tests if the communicator is an inter-communicator.

Standard sections(s)

6.6.1, See MPI_Comm_test_inter

fn set_name(&self, name: &str)

Set the communicator name

Standard section(s)

6.8, see the MPI_Comm_set_name function

fn get_name(&self) -> String

Get the communicator name

Standard section(s)

6.8, see the MPI_Comm_get_name function

fn create_cartesian_communicator( &self, dims: &[Count], periods: &[bool], reorder: bool ) -> Option<CartesianCommunicator>

Creates a communicator with ranks laid out in a multi-dimensional space, allowing for easy neighbor-to-neighbor communication, while providing MPI with information to allow it to better optimize the physical locality of ranks that are logically close.

dims - array of spatial extents for the cartesian space

periods - Must match length of dims. For i in 0 to dims.len(), periods[i] indicates if axis i is periodic. i.e. if true, the element at dims[i] - 1 in axis i is a neighbor of element 0 in axis i

reorder - If true, MPI may re-order ranks in the new communicator.

Standard section(s)

7.5.1 (MPI_Cart_create)

fn cartesian_map(&self, dims: &[Count], periods: &[bool]) -> Option<Rank>

Gets the target rank of this rank as-if create_cartesian_communicator had been called with dims, periods, and reorder = true.

Returns None if the local process would not particate in the new CartesianCommunciator.

dims - array of spatial extents for the cartesian space

periods - Must match length of dims. For i in 0 to dims.len(), periods[i] indicates if axis i is periodic. i.e. if true, the element at dims[i] - 1 in axis i is a neighbor of element 0 in axis i

Standard section

7.5.8 (MPI_Cart_map)

fn pack_size<Dt>(&self, incount: Count, datatype: &Dt) -> Countwhere Dt: Datatype, Self: Sized,

Gets the implementation-defined buffer size required to pack ‘incount’ elements of type ‘datatype’.

Standard section(s)

4.2, see MPI_Pack_size

fn pack<Buf>(&self, inbuf: &Buf) -> Vec<u8>where Buf: ?Sized + Buffer, Self: Sized,

Packs inbuf into a byte array with an implementation-defined format. Often paired with unpack to convert back into a specific datatype.

Standard Sections

4.2, see MPI_Pack

fn pack_into<Buf>( &self, inbuf: &Buf, outbuf: &mut [u8], position: Count ) -> Countwhere Buf: ?Sized + Buffer, Self: Sized,

Packs inbuf into a byte array with an implementation-defined format. Often paired with unpack to convert back into a specific datatype.

Standard Sections

4.2, see MPI_Pack

unsafe fn unpack_into<Buf>( &self, inbuf: &[u8], outbuf: &mut Buf, position: Count ) -> Countwhere Buf: ?Sized + BufferMut, Self: Sized,

Unpacks an implementation-specific byte array from pack or pack_into into a buffer of a specific datatype.

Standard Sections

4.2, see MPI_Unpack

fn parent(&self) -> Option<SystemInterCommunicator>

Returns the parent Communicator, if any

Standard Sections

10.3.2, see MPI_Comm_get_parent